+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # IntelliJ Markdown Changelog
 
 ## [Unreleased]
+- Prevent intersecting nodes for broken links
 
 ## [0.7.10]
 - [#211] Add support for the GFM tagfilter extension